[metrics] Fix metrics unwrap
